requirements.txt
requests>=2.31.0
yt-dlp>=2024.1.0
openai-whisper>=20231117
